requirements.txt
fastapi
numpy
Pillow
requests
python-multipart
uvicorn
tensorflow
tensorflow-hub